+++ src/config.py
@@ -6,7 +6,7 @@ from dotenv import load_dotenv
 load_dotenv()
 
 # --- System Version ---
-VERSION = "13.1.3"
+VERSION = "13.1.4"
 
 def load_config():
     """